+++ test_init_prio_b.cpp
@@ -1,5 +1,5 @@
 #include "test_init_prio.hpp"
 
-int b = a;
+float b = a;
